@@ -1,18 +1,18 @@
 name: qiniu_sdk_base
-version: 0.2.2
+version: 0.3.0
 homepage: https://github.com/qiniu/dart-sdk
 description: The sdk basic of Qiniu products
 
 environment:
-  sdk: '>=2.7.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  dio: ^3.0.10
-  crypto: ^2.1.5
-  meta: ^1.2.3
+  dio: ^4.0.0
+  crypto: ^3.0.0
+  meta: ^1.3.0
 
 dev_dependencies:
-  pedantic: ^1.9.0
-  test: ^1.14.4
-  dotenv: ^1.0.0
-  test_coverage: "^0.5.0"
+  pedantic: ^1.11.0
+  test: ^1.16.8
+  dotenv: 3.0.0-nullsafety.0
+  coverage: "^1.0.2"
